refactor(settings): 将个人中心相关类型定义抽离至独立的 types 目录
